# openapi.json
    {"openapi":"3.1.0","info":{"title":"Traffic Analytics API","description":"API for traffic and weather data analytics","version":"1.0.0"},"paths":{"/api/v1/data/data/refresh":{"post":{"tags":["Refresh","Refresh"],"summary":"Refresh Data","description":"Trigger Google extractor + weather extractor + ETL loaders for given city and lat/lon.\nSynchronous — returns summary after ETL completes.","operationId":"refresh_data_api_v1_data_data_refresh_post","parameters":[{"name":"city","in":"query","required":true,"schema":{"type":"string","title":"City"}},{"name":"lat","in":"query","required":true,"schema":{"type":"number","title":"Lat"}},{"name":"lon","in":"query","required":true,"schema":{"type":"number","title":"Lon"}},{"name":"headless","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Headless"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/traffic/summary":{"get":{"tags":["Traffic","Traffic"],"summary":"Get Summary","operationId":"get_summary_api_v1_traffic_summary_get","parameters":[{"name":"city","in":"query","required":true,"schema":{"type":"string","title":"City"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/traffic/history":{"get":{"tags":["Traffic","Traffic"],"summary":"Get History","operationId":"get_history_api_v1_traffic_history_get","parameters":[{"name":"city","in":"query","required":true,"schema":{"type":"string","title":"City"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/weather/current":{"get":{"tags":["Weather","Weather"],"summary":"Get Current","operationId":"get_current_api_v1_weather_current_get","parameters":[{"name":"city","in":"query","required":true,"schema":{"type":"string","title":"City"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}